Rename sketch-fab to sketchfab
